Add Redis data storage plugin
Any data for the bot should now be saved to a Redis store so there
should not be a requirement for data to be written directly to the
disk/machine that it is running on.

It seems that this bot framework expects an admin user to install
plugins directly through the chat interface:

`repos install <plugin-github-url>`

However this means that none of the plugins are in source control and
would have to be manually reinstalled if the bot ever had to be rebuilt.

It's not ideal to check in the plugins, especially not doing so as a
submodule, but it does mean that we can rebuild everything if needs be
from scratch at any time.